Feature: support gitlab urls with folders in repo
Gitlab urls may contain more than a folder in the repo name.
So, setting the clone location properly in the object and
creating the proper subdirectory for github folders.

https://gitlab.com/supercash/services/sms-gateway-service

In this case, services/sms is the repo part as part of the
regex validating the URL. The dir services is created.