:art: Isolate CLI implementation to be testable
The current code couldn't be tested because of the basic implementation using cobra.
This commit refacts the implementation to isolate the interface struct so that
we can have a better reuse of the output of the commands. This is based on the
implementation of the Docker CLI structure described at the following:

https://stackoverflow.com/questions/59709345/how-to-implement-unit-tests-for-cli-commands-in-go/59714127#59714127