:recycle: Split CMD implementation as a Controller
The isolated method GitCloneCmd now can be tested in isolation,
since it can be placed properly in test files. Note that there's
a clear separation of Cobra and the Service implementation.

The Cmd implementation is now moved to the specific API git controller,
which implements the entire execution and returns necessary status
code and errors as part of the API contract.