:sparkles: Add private key ssh handler: solves #14
As the format of git@ URLs use SSH keys for cloning. It requires
the use of the auth tokens based on the keys. A default key
is provided as usual, but the user should be able to provide with
an option.

https://github.com/src-d/go-git/issues/550#issuecomment-543013929

The error was first described at https://github.com/go-git/go-git/issues/169
and on #14.