Makefile & github flow: executing tests for #18
Moving the decrypt function to Makefile so that we can
run the tests locally if needed.

Updating the .gitignore with the paths to not use

Moved the pgp file into the scripts dir for testing

This should solve Issue #18 that requires tests using
private keys, so I extended it to use the docker version.