:children_crossing: Repo param is now optional and default
The verification of the repo param is now default. The solution
described at https://github.com/spf13/cobra/issues/818#issuecomment-489021216
requires the value to be retrieved by the args. That way, we need
to first verify if it is available.

The structure remains the same, but the param can be used with the
values as well.